# Improved performance of first enumeration