Improved the performance of Efficient GCF Enumeration and implemented a new enumeration called Parallel GCF Enumeration. The improvement to the efficient algorithm is usage of infinite precision integer math instead of mpmath floats. The parallel implementation uses numpy matrix math, but still uses only one core. Instead, it uses additional memory, which currently is limited to 0.1GB (10GB had only marginally better performance, i.e. ~1%). 

Benchmarks shows that the improved Efficient GCF Enumeration has an execution time that is about half of the original. The Parallel GCF Enumeration further reduces the execution time to less than a third of the improved version, for a total run time of less than 15% of the original algorithm. 

Details: 
e depth 20. Cartesian polynomial with an of order 1 in range [-20, 20] and bn of order 2 in range [-20, 20]. Execution time of whole script (with e pre-calculated): parallel: 146s; improved efficient: 599s; old efficient: 1256s.
pi depth 20. Cartesian polynomial with an of order 1 in range [-5, 5] and bn of order 2 in range [-30, 30]. Execution time of whole script (with pi pre-calculated): parallel: 37s; improved efficient: 125s; old efficient: 255s.

I've also added a prediction of the execution time left for the heaviest step (first enumeration) to the print messages. 

Please put it through its paces on your own searches! Any feedback is welcome.